Fixed the following error when building an hosting app:
```
TypeError: Arguments to path.join must be strings
    at Object.posix.join (path.js:488:13)
    at tryFile (/path/to/app-name/node_modules/addon-name/node_modules/ember-cli-sass/index.js:43:25)
    at /path/to/app-name/node_modules/addon-name/node_modules/ember-cli-sass/index.js:26:15
```

The root of the error is that `ember-cli` passes different `tree` as the
parameter to `toTree()`.

When building an addon that directly uses `ember-cli-sass`, `tree` is a `String`
that contains `"/path/to/your/addon-name/addon/styles"`.

When building an hosting app that indirectly uses `ember-cli-sass` through an
addon, `tree` is an object with a `read()` function that returns
`"/path/to/your/addon-name/addon/styles"`.